backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.3
tensorflow==2.15.0
pillow==10.2.0
